+++ Forms/LogForm.vb
@@ -13,12 +13,16 @@ Public Class LogForm
     End Sub
 
     Sub Init()
-        rtb.Text = File.ReadAllText(Path)
+        rtb.Text = Path.ReadAllText
 
         For Each match As Match In Regex.Matches(rtb.Text, "----- (.+) -----")
             Dim val = match.Groups(1).Value
             Dim match2 = Regex.Match(val, " \d+\.+.+")
-            If match2.Success Then val = val.Substring(0, val.Length - match2.Value.Length)
+
+            If match2.Success Then
+                val = val.Substring(0, val.Length - match2.Value.Length)
+            End If
+
             lb.Items.Add(val)
         Next